1. add a CANID into the CANOBD_Message_Parameter_Table. The CANID will be used as the PGNDA00. 2. in order to match the extended frame, the CAN_message_ID was re-defined as uint32_t

+++ LLD/Communications/OBD/CANIO/cn_io_transfer.h
@@ -45,7 +45,7 @@ typedef enum
 
 typedef struct CAN_Message_Parameter_Tag
 {
-    uint16_t                      CAN_message_ID;
+    uint32_t                      CAN_message_ID;
     uint8_t                       CAN_message_length;
     uint8_t                       CAN_buffer_depth;
     uint8_t                       *CAN_buffer_ptr;
